+++ Source/SwiftLintFramework/Rules/LineLengthRule.swift
@@ -18,15 +18,19 @@ public struct LineLengthRule: ConfigurationProviderRule, SourceKitFreeRule {
         name: "Line Length",
         description: "Lines should not span too many characters.",
         nonTriggeringExamples: [
-            Repeat(count: 100, repeatedValue: "/").joinWithSeparator("") + "\n"
+            Repeat(count: 100, repeatedValue: "/").joinWithSeparator("") + "\n",
+            Repeat(count: 100, repeatedValue: "#colorLiteral(red: 0.9607843161, green: 0.7058823705, blue: 0.200000003, alpha: 1)").joinWithSeparator("") + "\n",
+            Repeat(count: 100, repeatedValue: "#imageLiteral(resourceName: \"image.jpg\")").joinWithSeparator("") + "\n"
         ],
         triggeringExamples: [
-            Repeat(count: 101, repeatedValue: "/").joinWithSeparator("") + "\n"
+            Repeat(count: 101, repeatedValue: "/").joinWithSeparator("") + "\n",
+            Repeat(count: 101, repeatedValue: "#colorLiteral(red: 0.9607843161, green: 0.7058823705, blue: 0.200000003, alpha: 1)").joinWithSeparator("") + "\n",
+            Repeat(count: 101, repeatedValue: "#imageLiteral(resourceName: \"image.jpg\")").joinWithSeparator("") + "\n"
         ]
     )
 
     public func validateFile(file: File) -> [StyleViolation] {
-        let minValue = configuration.params.map({$0.value}).minElement(<)
+        let minValue = configuration.params.map({ $0.value }).minElement(<)
         return file.lines.flatMap { line in
             // `line.content.characters.count` <= `line.range.length` is true.
             // So, `check line.range.length` is larger than minimum parameter value.
@@ -34,7 +38,15 @@ public struct LineLengthRule: ConfigurationProviderRule, SourceKitFreeRule {
             if line.range.length < minValue {
                 return nil
             }
-            let length = line.content.characters.count
+
+            var strippedString = line.content
+            strippedString = stripLiterals(fromSourceString: strippedString,
+                withDelimiter: "#colorLiteral")
+            strippedString = stripLiterals(fromSourceString: strippedString,
+                withDelimiter: "#imageLiteral")
+
+            let length = strippedString.characters.count
+
             for param in configuration.params where length > param.value {
                 return StyleViolation(ruleDescription: self.dynamicType.description,
                     severity: param.severity,
@@ -45,4 +57,34 @@ public struct LineLengthRule: ConfigurationProviderRule, SourceKitFreeRule {
             return nil
         }
     }
+
+    /// Takes a string and replaces any literals specified by the `delimiter` parameter with `#`
+    ///
+    /// - parameter sourceString: Original string, possibly containing literals
+    /// - parameter delimiter:    Delimiter of the literal
+    ///     (characters before the parentheses, e.g. `#colorLiteral`)
+    ///
+    /// - returns: sourceString with the given literals replaced by `#`
+    private func stripLiterals(fromSourceString sourceString: String,
+                                                withDelimiter delimiter: String) -> String {
+        var modifiedString = sourceString
+
+        // While copy of content contains literal, replace with a single character
+        while modifiedString.containsString("\(delimiter)(") {
+            if let rangeStart = modifiedString.rangeOfString("\(delimiter)("),
+                let rangeEnd = modifiedString.rangeOfString(")",
+                                                            options: .LiteralSearch,
+                                                            range:
+                    rangeStart.startIndex..<modifiedString.endIndex,
+                                                            locale: nil) {
+                modifiedString.replaceRange(rangeStart.startIndex..<rangeEnd.endIndex, with: "#")
+
+            } else { // Should never be the case, but break to avoid accidental infinity loop
+                break
+            }
+        }
+
+        return modifiedString
+    }
+
 }