+++ Source/SwiftLintFramework/Rules/RedundantNilCoalescingRule.swift
@@ -1,5 +1,5 @@
 //
-//  RedundantNilCoalesingRule.swift
+//  RedundantNilCoalescingRule.swift
 //  SwiftLint
 //
 //  Created by Daniel Beard on 8/24/16.
@@ -10,7 +10,7 @@ import Foundation
 import SourceKittenFramework
 
 extension File {
-    private func violatingRedundantNilCoalesingRanges() -> [NSRange] {
+    private func violatingRedundantNilCoalescingRanges() -> [NSRange] {
         return matchPattern(
             "\\?\\?\\s*nil\\b", // ?? {whitespace} nil {word boundary}
             excludingSyntaxKinds: SyntaxKind.commentAndStringKinds()
@@ -18,17 +18,17 @@ extension File {
     }
 }
 
-public struct RedundantNilCoalesingRule: OptInRule, ConfigurationProviderRule {
+public struct RedundantNilCoalescingRule: OptInRule, ConfigurationProviderRule {
 
     public var configuration = SeverityConfiguration(.Warning)
 
     public init() {}
 
     public static let description = RuleDescription(
-        identifier: "redundant_nil_coalesing",
-        name: "Redundant Nil Coalesing",
+        identifier: "redundant_nil_coalescing",
+        name: "Redundant Nil Coalescing",
         description: "nil coalescing operator is only evaluated if the lhs is nil " +
-            ", coalesing operator with nil as rhs is redundant",
+            ", coalescing operator with nil as rhs is redundant",
         nonTriggeringExamples: [
             "var myVar: Int?; myVar ?? 0"
         ],
@@ -38,7 +38,7 @@ public struct RedundantNilCoalesingRule: OptInRule, ConfigurationProviderRule {
     )
 
     public func validateFile(file: File) -> [StyleViolation] {
-        return file.violatingRedundantNilCoalesingRanges().map {
+        return file.violatingRedundantNilCoalescingRanges().map {
             StyleViolation(ruleDescription: self.dynamicType.description,
                 severity: configuration.severity,
                 location: Location(file: file, characterOffset: $0.location))