+++ Source/SwiftLintFramework/Rules/RuleConfigurations/TrailingWhitespaceConfiguration.swift
@@ -11,14 +11,17 @@ import Foundation
 public struct TrailingWhitespaceConfiguration: RuleConfiguration, Equatable {
     var severityConfiguration = SeverityConfiguration(.Warning)
     var ignoresEmptyLines = false
+    var ignoresComments = true
 
     public var consoleDescription: String {
         return severityConfiguration.consoleDescription +
-            ", ignores_empty_lines: \(ignoresEmptyLines)"
+            ", ignores_empty_lines: \(ignoresEmptyLines)" +
+            ", ignores_comments: \(ignoresComments)"
     }
 
-    public init(ignoresEmptyLines: Bool) {
+    public init(ignoresEmptyLines: Bool, ignoresComments: Bool) {
         self.ignoresEmptyLines = ignoresEmptyLines
+        self.ignoresComments = ignoresComments
     }
 
     public mutating func applyConfiguration(configuration: AnyObject) throws {
@@ -27,6 +30,7 @@ public struct TrailingWhitespaceConfiguration: RuleConfiguration, Equatable {
         }
 
         ignoresEmptyLines = (configuration["ignores_empty_lines"] as? Bool == true)
+        ignoresComments = (configuration["ignores_comments"] as? Bool == true)
 
         if let severityString = configuration["severity"] as? String {
             try severityConfiguration.applyConfiguration(severityString)
@@ -36,5 +40,6 @@ public struct TrailingWhitespaceConfiguration: RuleConfiguration, Equatable {
 
 public func == (lhs: TrailingWhitespaceConfiguration,
                 rhs: TrailingWhitespaceConfiguration) -> Bool {
-    return lhs.ignoresEmptyLines == rhs.ignoresEmptyLines
+    return lhs.ignoresEmptyLines == rhs.ignoresEmptyLines &&
+        lhs.ignoresComments == rhs.ignoresComments
 }