+++ CHANGELOG.md
@@ -6,12 +6,70 @@
 
 ##### Enhancements
 
+* None.
+
+##### Bug Fixes
+
+* None.
+
+## 0.13.0: MakeYourClothesCleanAgain
+
+##### Breaking
+
+* None.
+
+##### Enhancements
+
+* Add `ignores_comment` configuration for `trailing_whitespace` rule.  
+  [Javier Hernández](https://github.com/jaherhi)
+  [#576](https://github.com/realm/SwiftLint/issues/576)
+
+* Added HTML reporter, identifier is `html`.  
+  [Johnykutty Mathew](https://github.com/Johnykutty)
+
+* Add `SuperCallRule` Opt-In rule that warns about methods not calling to super.  
+  [Angel G. Olloqui](https://github.com/angelolloqui)
+  [#803](https://github.com/realm/SwiftLint/issues/803)
+
 * Add `RedundantNilCoalesingRule` Opt-In rule that warns against `?? nil`.  
   [Daniel Beard](https://github.com/daniel-beard)
   [#764](https://github.com/realm/SwiftLint/issues/764)
 
+* Added opt-In rule to makes closure expressions spacing consistent.  
+  [J. Cheyo Jimenez](https://github.com/masters3d)
+  [#770](https://github.com/realm/SwiftLint/issues/770)
+
 * Adds `allow_private_set` configuration for the `private_outlet` rule.  
   [Rohan Dhaimade](https://github.com/HaloZero)
+
+* Swift 2.3 support.  
+  [Norio Nomura](https://github.com/norio-nomura),
+  [Syo Ikeda](https://github.com/ikesyo)
+
+* Color literals count as single characters to avoid unintentional line length
+  violations.  
+  [Jonas](https://github.com/VFUC)
+  [#742](https://github.com/realm/SwiftLint/issues/742)
+
+* Add `SwitchCaseOnNewlineRule` opt-in rule that enforces a newline after
+  `case pattern:` in a `switch`.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#681](https://github.com/realm/SwiftLint/issues/681)
+
+* Add `ValidIBInspectableRule` rule that checks if `@IBInspectable` declarations
+  are valid. An `@IBInspectable` is valid if:
+  * It's declared as a `var` (not `let`)
+  * Its type is explicit (not inferred)
+  * Its type is one of the
+  [supported types](http://help.apple.com/xcode/mac/8.0/#/devf60c1c514)  
+
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#756](https://github.com/realm/SwiftLint/issues/756)  
+
+* Add ``ExplicitInitRule`` Opt-In rule to discourage calling ``init``
+  directly.  
+  [Matt Taube](https://github.com/mtaube)
+  [#715](https://github.com/realm/SwiftLint/pull/715)
 
 ##### Bug Fixes
 
@@ -24,6 +82,10 @@
   [J. Cheyo Jimenez](https://github.com/masters3d)
   [#772](https://github.com/realm/SwiftLint/issues/772)
 
+* Correctable rules no longer apply corrections if the rule is locally disabled.  
+  [J. Cheyo Jimenez](https://github.com/masters3d)  
+  [#601](https://github.com/realm/SwiftLint/issues/601)
+
 * Fixed regex bug in Mark Rule where MARK could not be used with only a hyphen
   but no descriptive text: `// MARK: -`.  
   [Ruotger Deecke](https://github.com/roddi)
@@ -33,6 +95,20 @@
   Fixed: Private unit test rule config is ignored if regex is missing.  
   [Cristian Filipov](https://github.com/cfilipov)
   [#786](https://github.com/realm/SwiftLint/issues/786)
+
+* Fixed: `ConditionalReturnsOnNewline` now respects severity configuration.  
+  [Rohan Dhaimade](https://github.com/HaloZero)
+  [#783](https://github.com/realm/SwiftLint/issues/783)
+
+* Fixed: `ConditionalReturnsOnNewline` now checks if `return` is a keyword,
+  avoiding false positives.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#784](https://github.com/realm/SwiftLint/issues/784)
+
+* `ForceUnwrappingRule` did not recognize force unwraps in return statements
+  using subscript.  
+  [Norio Nomura](https://github.com/norio-nomura)
+  [#813](https://github.com/realm/SwiftLint/issues/813)  
 
 ## 0.12.0: Vertical Laundry
 
@@ -101,7 +177,7 @@ This release has seen a phenomenal uptake in community contributions!
   [bootstraponline](https://github.com/bootstraponline)
   [#689](https://github.com/realm/SwiftLint/issues/689)
 
-* Add rule to check for private unit tests (private unit tests don't get run 
+* Add rule to check for private unit tests (private unit tests don't get run
   by XCTest).  
   [Cristian Filipov](https://github.com/cfilipov)
 
@@ -111,7 +187,7 @@ This release has seen a phenomenal uptake in community contributions!
 
 * Adds 'ConditionalReturnsOnNewLineRule' rule.  
   [Rohan Dhaimade](https://github.com/HaloZero)
-  
+
 * Made `- returns:` doc optional for initializers.  
   [Mohpor](https://github.com/mohpor)
   [#557](https://github.com/realm/SwiftLint/issues/557)